refactor: move files to its directories
