+++ src/components/AppStateContext.tsx
@@ -1,0 +1,44 @@
+import React, { useReducer } from "react";
+import { createContext } from "react";
+import { Action } from "../actions/types";
+import { appStateReducer } from "../reducers/appState";
+import { AppState } from "../reducers/types";
+
+const appData: AppState = {
+  draggedItem: undefined,
+  lists: [
+    {
+      id: "0",
+      text: "To Do",
+      tasks: [{ id: "c0", text: "Generate app scaffold" }],
+    },
+    {
+      id: "1",
+      text: "In Progress",
+      tasks: [{ id: "c2", text: "Learn Typescript" }],
+    },
+    {
+      id: "2",
+      text: "Done",
+      tasks: [{ id: "c3", text: "Begin to use static typing" }],
+    },
+  ],
+};
+
+interface AppStateContextProps {
+  state: AppState;
+  dispatch: React.Dispatch<Action>;
+}
+
+export const AppStateContext = createContext<AppStateContextProps>(
+  {} as AppStateContextProps
+);
+
+export const AppStateProvider = ({ children }: React.PropsWithChildren<{}>) => {
+  const [state, dispatch] = useReducer(appStateReducer, appData);
+  return (
+    <AppStateContext.Provider value={{ state, dispatch }}>
+      {children}
+    </AppStateContext.Provider>
+  );
+};
