+++ src/components/NewItemForm.tsx
@@ -1,7 +1,11 @@
 import React, { useState } from "react";
 
-import { NewItemButton, NewItemFormContainer, NewItemInput } from "./styles";
-import { useFocus } from "./hooks/useFocus";
+import {
+  NewItemButton,
+  NewItemFormContainer,
+  NewItemInput,
+} from "../styles/styles";
+import { useFocus } from "../hooks/useFocus";
 
 interface NewItemFormProps {
   onAdd(text: string): void;
