+++ src/hooks/useItemDrag.ts
@@ -1,8 +1,9 @@
 import { useDrag } from "react-dnd";
 import { useAppState } from "./useAppState";
-import { DragItem } from "../DragItem";
+import { DragItem } from "../dragAndDrop/DragItem";
 import { useEffect } from "react";
 import { getEmptyImage } from "react-dnd-html5-backend";
+import { ActionTypes } from "../actions/types";
 
 export const useItemDrag = (item: DragItem) => {
   const { dispatch } = useAppState();
@@ -10,13 +11,13 @@ export const useItemDrag = (item: DragItem) => {
     item,
     begin: () => {
       dispatch({
-        type: "SET_DRAGGED_ITEM",
+        type: ActionTypes.setDraggedItem,
         payload: item,
       });
     },
     end: () => {
       dispatch({
-        type: "SET_DRAGGED_ITEM",
+        type: ActionTypes.setDraggedItem,
         payload: undefined,
       });
     },
