+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import { AppStateProvider } from "./AppStateContext";
+import "./styles/index.css";
+import App from "./components/App";
+import { AppStateProvider } from "./components/AppStateContext";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend as Backend } from "react-dnd-html5-backend";
