+++ src/reducers/appState.ts
@@ -1,0 +1,63 @@
+import { Action, ActionTypes } from "../actions/types";
+import { nanoid } from "nanoid";
+import { findItemIndexById } from "../utils/findItemIndexById";
+import { moveItem } from "../utils/moveItem";
+import { AppState } from "./types";
+
+export const appStateReducer = (state: AppState, action: Action): AppState => {
+  switch (action.type) {
+    case ActionTypes.addList: {
+      return {
+        ...state,
+        lists: [
+          ...state.lists,
+          { id: nanoid(), text: action.payload, tasks: [] },
+        ],
+      };
+    }
+    case ActionTypes.addTask: {
+      const targetLaneIndex = findItemIndexById(
+        state.lists,
+        action.payload.listId
+      );
+      state.lists[targetLaneIndex].tasks.push({
+        id: nanoid(),
+        text: action.payload.text,
+      });
+      return {
+        ...state,
+      };
+    }
+    case ActionTypes.moveList: {
+      const { dragIndex, hoverIndex } = action.payload;
+      state.lists = moveItem(state.lists, dragIndex, hoverIndex);
+      return {
+        ...state,
+      };
+    }
+
+    case ActionTypes.setDraggedItem: {
+      return { ...state, draggedItem: action.payload };
+    }
+
+    case ActionTypes.moveTask: {
+      const {
+        dragIndex,
+        hoverIndex,
+        sourceColumn,
+        targetColumn,
+      } = action.payload;
+
+      const sourceLaneIndex = findItemIndexById(state.lists, sourceColumn);
+      const targetLaneIndex = findItemIndexById(state.lists, targetColumn);
+
+      const item = state.lists[sourceLaneIndex].tasks.splice(dragIndex, 1)[0];
+      state.lists[targetLaneIndex].tasks.splice(hoverIndex, 0, item);
+
+      return { ...state };
+    }
+
+    default:
+      return state;
+  }
+};
